+++ src/3btree/btree_records_pod.h
@@ -93,7 +93,7 @@ class PodRecordList : public BaseRecordList
     }
 
     // Returns the record size
-    uint64_t get_record_size(Context *context, int slot,
+    uint32_t get_record_size(Context *context, int slot,
                     int duplicate_index = 0) const {
       return (sizeof(PodType));
     }
