+++ src/4cursor/cursor_local.h
@@ -336,7 +336,7 @@ class LocalCursor : public Cursor
                                 uint32_t *pcount);
 
     // Get current record size (ups_cursor_get_record_size)
-    virtual ups_status_t do_get_record_size(uint64_t *psize);
+    virtual ups_status_t do_get_record_size(uint32_t *psize);
 
     // Implementation of get_duplicate_position()
     virtual ups_status_t do_get_duplicate_position(uint32_t *pposition);
